changed display format for default power on lcd screen

+++ Kcore_103/KCore/adc/adc.h
@@ -29,7 +29,7 @@ typedef struct {
 	uint16_t    adcRaw; // last raw value read from ADC
 	int16_t     convRaw; // converted value from last rawValue read
 	int16_t     adcAvg; // average of the last 10 read (or converted reads) after tossing high and low
-	int16_t     convAvg; // converted value using the rawAvg
+	float     convAvg; // converted value using the rawAvg
 	int16_t     sampleHistory[ADC_NUM_SAMPLES]; // last N reads from ADC
 	uint8_t     sampleIndex; // index of last valid value in the history
 	uint8_t     inputChannel; // actual adc channel input selection
